AI/Text/requirements.txt
python-dotenv
langdetect
langchain-core
langchain-huggingface
huggingface_hub
transformers
argparse
